Fix issues with tests

+++ nebullvm/operations/optimizations/tests/test_openvino.py
@@ -164,7 +164,9 @@ def test_openvino(
             res_orig = tuple(model(*inputs_example))
             assert all(
                 [
-                    torch.allclose(res_tensor, res_orig_tensor, rtol=1e-01)
+                    torch.allclose(
+                        res_tensor.float(), res_orig_tensor, rtol=1e-01
+                    )
                     for (res_tensor, res_orig_tensor) in zip(res, res_orig)
                 ]
             )
