+++ nebullvm/operations/optimizations/tests/test_tensor_rt.py
@@ -170,7 +170,9 @@ def test_tensorrt_onnx(
             res_orig = tuple(model(*inputs_example))
             assert all(
                 [
-                    torch.allclose(res_tensor, res_orig_tensor, rtol=1e-01)
+                    torch.allclose(
+                        res_tensor.float(), res_orig_tensor, rtol=1e-01
+                    )
                     for (res_tensor, res_orig_tensor) in zip(res, res_orig)
                 ]
             )
@@ -298,7 +300,9 @@ def test_tensorrt_torch(
             res_orig = tuple(model(*inputs_example))
             assert all(
                 [
-                    torch.allclose(res_tensor, res_orig_tensor, rtol=1e-01)
+                    torch.allclose(
+                        res_tensor.float(), res_orig_tensor, rtol=1e-01
+                    )
                     for (res_tensor, res_orig_tensor) in zip(res, res_orig)
                 ]
             )
