+++ nebullvm/operations/optimizations/tests/test_torchscript.py
@@ -118,7 +118,9 @@ def run_test_torchscript(
             res_orig = tuple(model(*inputs_example))
             assert all(
                 [
-                    torch.allclose(res_tensor, res_orig_tensor, rtol=1e-01)
+                    torch.allclose(
+                        res_tensor.float(), res_orig_tensor, rtol=1e-01
+                    )
                     for (res_tensor, res_orig_tensor) in zip(res, res_orig)
                 ]
             )
